RavenDB-18756 implemented /databases/*/multi_get endpoint
